requirements.txt
PyPDF2
python-docx
python-pptx
Pillow
pytesseract
gTTS
opencv-python
numpy